Add day 14
This one was fun to do, because I've tried to use `comptime` and
some bit twiddling hacks. Unfortunately there is a possible bug in `std.HashMap`
and codegen when using `inline while` with `comptime` vars.
The solution can be optimized further by preprocessing `mask` ranges in
part 2, but it is a bit time-consuming.