Add day 17
This one was pretty interesting because of compiletime usage.
Compilation hung with my elegant first solution (n-dimensional slice as
a structure with methods), so I've chosen to use raw slices and
package-level functions. This solution is much less elegant but can be
easily reused for bigger dimensions. The goal of this exercise to me was
to make both part 1 and 2 work via a single algorithm with a
compile-time dimension parameter.